pyproject.toml
[project]
name = "northwind-streamlit-migration"
version = "0.1.0"
description = "Northwind PowerBI to Streamlit Migration"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "numpy>=1.24.0",
    "snowflake-connector-python>=3.0.0",
    "psycopg2-binary>=2.9.0",
    "cryptography>=41.0.0",
]

[tool.uv]
dev-dependencies = []
